Fix decision_matrix log setup
The decision_matrix log was using chunk_size of 1 per scenario which
results in a very large memmory allocation and was leading to a very
long time for result set setup. This commit makes change the chunk_size
of decision_matrix and make other modifications to follow the same
pattern as other logs.

+++ src/scenario.jl
@@ -491,7 +491,7 @@ function _write_batch!(
     # decision_matrix_log: (tf, n_locs, n_criteria, n)
     if parse(Bool, get(ENV, "ADRIA_LOG_DM", "false")) == true
         _, n_l, n_c = size(results[1].decision_matrix_log)
-        dm_batch = Array{Float64}(undef, tf, n_l, n_c, n)
+        dm_batch = Array{Float16}(undef, tf, n_l, n_c, n)
         for (i, r) in enumerate(results)
             dm_batch[:, :, :, i] .= r.decision_matrix_log
         end
@@ -1136,7 +1136,7 @@ function run_model(
 
     # Decision matrix log
     if log_dm
-        decision_matrix_log = ZeroDataCube(; T=Float64, timesteps=1:tf,
+        decision_matrix_log = ZeroDataCube(; T=Float32, timesteps=1:tf,
             location=domain.loc_ids[habitable_locs], criteria=seed_pref.names)
     else
         decision_matrix_log = false
